# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Technician Dispatch & Service Tracking System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  customer: "customer",
  technician: "technician",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  phone?: string;
  role?: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  technician: "technician",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  phone?: string;
  isActive: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateUserBody {
  name?: string;
  phone?: string;
  isActive?: boolean;
}

export interface TechnicianProfile {
  id: number;
  name: string;
  email: string;
  phone?: string;
  isActive: boolean;
  totalJobs: number;
  activeJobs: number;
}

export type ServiceRequestServiceType =
  (typeof ServiceRequestServiceType)[keyof typeof ServiceRequestServiceType];

export const ServiceRequestServiceType = {
  installation: "installation",
  repair: "repair",
  maintenance: "maintenance",
  amc_service: "amc_service",
  inspection: "inspection",
} as const;

export type ServiceRequestStatus =
  (typeof ServiceRequestStatus)[keyof typeof ServiceRequestStatus];

export const ServiceRequestStatus = {
  pending: "pending",
  assigned: "assigned",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ServiceRequestPriority =
  (typeof ServiceRequestPriority)[keyof typeof ServiceRequestPriority];

export const ServiceRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface ServiceRequest {
  id: number;
  title: string;
  description: string;
  serviceType: ServiceRequestServiceType;
  status: ServiceRequestStatus;
  priority: ServiceRequestPriority;
  customerId: number;
  technicianId?: number | null;
  customerName: string;
  technicianName?: string | null;
  address: string;
  technicianNotes?: string | null;
  scheduledAt?: string | null;
  completedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateServiceRequestBodyServiceType =
  (typeof CreateServiceRequestBodyServiceType)[keyof typeof CreateServiceRequestBodyServiceType];

export const CreateServiceRequestBodyServiceType = {
  installation: "installation",
  repair: "repair",
  maintenance: "maintenance",
  amc_service: "amc_service",
  inspection: "inspection",
} as const;

export type CreateServiceRequestBodyPriority =
  (typeof CreateServiceRequestBodyPriority)[keyof typeof CreateServiceRequestBodyPriority];

export const CreateServiceRequestBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface CreateServiceRequestBody {
  title: string;
  description: string;
  serviceType: CreateServiceRequestBodyServiceType;
  priority?: CreateServiceRequestBodyPriority;
  address: string;
  scheduledAt?: string;
}

export type UpdateServiceRequestBodyStatus =
  (typeof UpdateServiceRequestBodyStatus)[keyof typeof UpdateServiceRequestBodyStatus];

export const UpdateServiceRequestBodyStatus = {
  pending: "pending",
  assigned: "assigned",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type UpdateServiceRequestBodyPriority =
  (typeof UpdateServiceRequestBodyPriority)[keyof typeof UpdateServiceRequestBodyPriority];

export const UpdateServiceRequestBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface UpdateServiceRequestBody {
  status?: UpdateServiceRequestBodyStatus;
  technicianNotes?: string;
  scheduledAt?: string;
  priority?: UpdateServiceRequestBodyPriority;
}

export interface AssignTechnicianBody {
  technicianId: number;
}

export interface DashboardSummary {
  totalRequests: number;
  pendingRequests: number;
  inProgressRequests: number;
  completedRequests: number;
  totalTechnicians: number;
  totalCustomers: number;
  activeTechnicians: number;
}

export interface StatusBreakdown {
  pending: number;
  assigned: number;
  in_progress: number;
  completed: number;
  cancelled: number;
}

export type ListServiceRequestsParams = {
  status?: ListServiceRequestsStatus;
  technicianId?: number;
};

export type ListServiceRequestsStatus =
  (typeof ListServiceRequestsStatus)[keyof typeof ListServiceRequestsStatus];

export const ListServiceRequestsStatus = {
  pending: "pending",
  assigned: "assigned",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ListUsersParams = {
  role?: ListUsersRole;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  customer: "customer",
  technician: "technician",
  admin: "admin",
} as const;