# lib/api-zod/src/generated/types/assignTechnicianBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Technician Dispatch & Service Tracking System API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignTechnicianBody {
  technicianId: number;
}