# lib/api-zod/src/generated/types/registerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Technician Dispatch & Service Tracking System API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterBodyRole } from "./registerBodyRole";

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  phone?: string;
  role?: RegisterBodyRole;
}