# lib/api-zod/src/generated/types/serviceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Technician Dispatch & Service Tracking System API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceRequestPriority } from "./serviceRequestPriority";
import type { ServiceRequestServiceType } from "./serviceRequestServiceType";
import type { ServiceRequestStatus } from "./serviceRequestStatus";

export interface ServiceRequest {
  id: number;
  title: string;
  description: string;
  serviceType: ServiceRequestServiceType;
  status: ServiceRequestStatus;
  priority: ServiceRequestPriority;
  customerId: number;
  technicianId?: number | null;
  customerName: string;
  technicianName?: string | null;
  address: string;
  technicianNotes?: string | null;
  scheduledAt?: Date | null;
  completedAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}