# lib/api-zod/src/generated/types/statusBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Technician Dispatch & Service Tracking System API
 * OpenAPI spec version: 0.1.0
 */

export interface StatusBreakdown {
  pending: number;
  assigned: number;
  in_progress: number;
  completed: number;
  cancelled: number;
}